Got the GUI working
GUI works mostly fine.
GUI shows items in room and inventory as gibberish. This will need to be
solved somehow.

Not sure if Interceptor is properly implemented. Was talking with people
and they said it mightn't.

Also messed around with a few things. Changed most ArrayLists to
DefaultListModels because it works better with JList.